Lab 2 Pull Request (#1)
* Add Player Health Blueprint

This commit adds the health logic for the player:

-The player listens to when any damage is applied to it. If so, its health depletes until reaching 0. Upon reaching 0, it does a simple restart on the level.

-Two blueprint variables were created: max health and current health. max health is set to 100, and current health is set on initialization.

* Add Enemy Navmesh Pathfinding

This commit begins to add pathfinding logic for enemies.

Enemies have an enemy ai controller that can be customized to sense the player in a cone of view. If triggered, you can make the enemy follow the player until a new condition is met.

* Add Working Enemy Dealing Melee Damage

This commit adds the logic for the enemy dealing damage to the player through melee.

I used custom events (basically a method call) to help with logic flow. Using blueprint is very messy and I find it very difficult to follow the logic when it's scattered in 2 dimensions.

* Add Damage Visual Indicator

This commit adds a damage visual indicator for when the enemy hits the player. The player will have a particle system play, depicting blood coming out of them.

This commit should mark the completion of lab 2 requirements. To recap what was completed:

1)Enemy Behaviour for idling, detecting player, chasing the player, and attacking the player.
2) Use of state machines and custom events (methods) in blueprint to better control logic flow.
3) VFX for enemy hitting player

The rest TBD from GitHub itself.