Refactor Wave Management Pipeline
This commit does some refactoring to the wave management system to allow me to have greater control over what happens in the game. A box volume was added and connected to wave managers in the scene that will trigger it to start, an area clear manager will manage progression through the game, and the wave manager will send and receive events to manage the game state.