Add Wave Configuration
This commit adds the wave configuration for the game. There are a total of three wave areas with different wave combinations.

The next commit will aim to add player objective HUD and fix the wave manager to determine when the waves are truly complete (if time permits). Currently the wave goes on a timer to determine if the wave is complete, not checking if there are still enemies in the scene.